test/reftest/web-platform-tests/dom/nodes/getElementsByClassName-20.htm

<!DOCTYPE html>
<html><head>
        <title>getElementsByClassName</title>
        <meta content="get elements in document then add element to collection" name="description">
    <link href="https://dom.spec.whatwg.org/#dom-document-getelementsbyclassname" rel="help">
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
</head>
    <body>
        <div id="log"></div>
        <div>
            <div>
                <a class="text link" href="#foo">test link #foo</a>
            </div>
            <b class="text">text</b>
        </div>
        <!-- table-related tags are not supported
        <table>
            <caption class="text caption">text caption</caption>
            <thead>
                <tr>
                    <td class="TEXT head">TEXT head</td>
                </tr>
            </thead>
            <tbody>
                <tr>
                    <td class="td text1">td text1</td>
                </tr>
                <tr>
                    <td class="td text">td text</td>
                </tr>
                <tr>
                    <td class="td te xt">td te xt</td>
                </tr>
            </tbody>
            <tfoot>
                <tr>
                    <td class="TEXT foot">TEXT foot</td>
                </tr>
            </tfoot>
        </table>
        -->
        <div>
            <div class="text caption">text caption</div>
            <div>
                <div>
                    <span class="TEXT head">TEXT head</span>
                </div>
            </div>
            <div>
                <div>
                    <span class="span text1">span text1</span>
                </div>
                <div>
                    <span class="span text">span text</span>
                </div>
                <div>
                    <span class="span te xt">span te xt</span>
                </div>
            </div>
            <div>
                <div>
                    <span class="TEXT foot">TEXT foot</span>
                </div>
            </div>
        </div>
        <div class="xt te">xt te</div>

        <script type="text/javascript">
            test(function()
            {
                var collection = document.getElementsByClassName("text");
                assert_equals(collection.length, 4);
                var newDiv = document.createElement("div");
                newDiv.setAttribute("class", "text");
                // innerHTML is not supported
                // newDiv.innerHTML = "text newDiv";
                // document.getElementsByTagName("table")[0].tBodies[0].rows[0].cells[0].appendChild(newDiv);
                document.body.appendChild(newDiv);
                assert_equals(collection.length, 5);
                /*
                assert_equals(collection[0].parentNode.nodeName, "DIV");
                assert_equals(collection[1].parentNode.nodeName, "DIV");
                assert_equals(collection[2].parentNode.nodeName, "TABLE");
                assert_equals(collection[3].parentNode.nodeName, "TD");
                assert_equals(collection[4].parentNode.nodeName, "TR");
                */
                assert_equals(collection.item(0).parentNode.nodeName, "DIV");
                assert_equals(collection.item(1).parentNode.nodeName, "DIV");
                assert_equals(collection.item(2).parentNode.nodeName, "DIV");
                assert_equals(collection.item(3).parentNode.nodeName, "DIV");
                assert_equals(collection.item(4).parentNode.nodeName, "BODY");
            }, "get elements in document then add element to collection");
        </script>
</body></html>
